separate settings_json into it's own crate
