+++ crates/settings/Cargo.toml
@@ -35,6 +35,7 @@ serde_json_lenient.workspace = true
 serde_path_to_error.workspace = true
 serde_repr.workspace = true
 serde_with.workspace = true
+settings_json.workspace = true
 settings_macros.workspace = true
 smallvec.workspace = true
 strum.workspace = true
